boss scene change trigger and script
